CatsAndDogs: Remove unnecessary .to transformations
The .to transforms are a major contributor
to how long the training takes, so I decided
to at least get rid of some of the ones
that were still hanging around without being
particularly useful.

This did not make much of a difference, so
the real culprits are somewhere else, potentially
the ones that change to and from GPU in the forward
function. Another major contributor to the time
is the loss calculations. Both this and the
transformations seem to independently take roughly
the same time as the backward pass calculation.
In order to get rid of some of the transforms,
the calculations would have to be done more on
GPU in general. However, for this to be ideal,
the loops in the loss calculations would have
to be vectorised, as currently running the whole
thing on the GPU is very slow compared to
just the model on GPU and the rest on CPU.

+++ CatsAndDogs/src/utils/math.py
@@ -243,7 +243,7 @@ def symmetric_intersection(bb1: torch.Tensor, bb2: torch.Tensor) -> torch.Tensor
 
 
 
-def intersection_over_union(bb1: torch.Tensor, bb2: torch.Tensor, dtype = torch.float32) -> torch.Tensor:
+def intersection_over_union(bb1: torch.Tensor, bb2: torch.Tensor) -> torch.Tensor:
     '''
     Calculate the intersection-over-union AKA Jaccard index of
     the bounding boxes `bb1` (N,4), and `bb2`, (M,4). The format of a bounding
@@ -252,7 +252,10 @@ def intersection_over_union(bb1: torch.Tensor, bb2: torch.Tensor, dtype = torch.
     Return the values as (M,N).
     '''
 
-    intersections = symmetric_intersection(bb1.to(dtype), bb2.to(dtype))
+    intersections = symmetric_intersection(
+        bb1,
+        bb2
+    )
 
     bb1_areas = bounding_box_area(bb1)
     bb2_areas = bounding_box_area(bb2).reshape([-1,1])
@@ -260,13 +263,13 @@ def intersection_over_union(bb1: torch.Tensor, bb2: torch.Tensor, dtype = torch.
 
     return intersections/(areas_sum - intersections)
     
-def batched_intersection_over_union(bb1: torch.Tensor, bbs2: list[torch.Tensor], dtype = torch.float32) -> list[torch.Tensor]:
+def batched_intersection_over_union(bb1: torch.Tensor, bbs2: list[torch.Tensor]) -> list[torch.Tensor]:
     '''
     See `intersection_over_union()` for non-batched version. For bounding boxes
     per patch in `bbs2`, return a matching list of IoU values against
     `bb1`.
     '''
 
-    result = intersection_over_union(bb1, torch.vstack(bbs2), dtype = dtype)
+    result = intersection_over_union(bb1, torch.vstack(bbs2))
     idx = [0]+list(itertools.accumulate(map(len, bbs2)))
     return [result[start:end] for start,end in zip(idx[:-1], idx[1:])]
